+++ MOP/src/GameObjects/Items/UncleBeerCaseHook.cs
@@ -41,7 +41,7 @@ namespace MOP
                 return;
 
             // If Bottle Recycling mod is present, prevent destroying beer bottles
-            if (RuleFiles.instance.SpecialRules.DontDestroyEmptyBeerBottles)
+            if (Rules.instance.SpecialRules.DontDestroyEmptyBeerBottles)
             {
                 ModConsole.Print("<color=yellow>[MOP] Beer bottles won't be destroyed, because one or more mods prevent it. " +
                     "Disable 'Destroy empty beer bottles' in the MOP settings.</color>");